fast_api_service/requirements.txt
psycopg==3.1.12
psycopg_binary==3.1.12
psycopg2-binary==2.9.9
pandas==2.0.1
numpy==1.26.4
prometheus-fastapi-instrumentator==7.0.0
prometheus_client
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy
python-dotenv
python-multipart